+++ CHANGELOG.md
@@ -336,6 +336,7 @@
 * [X] 更改目录结构。
 * [X] breeze-dsl-command-line 优化代码。
 * [X] breeze-dsl-critic-markup & breeze-dsl-command-line 优化代码。
+* [X] breeze-core 添加`HandledCharSequence`，这个接口的输出字符串可能需要经过额外的处理。
 * [ ] breeze-core 让集合的`deepQuery`方法当特定的占位符不匹配类型时，回调为普通字符串。 
 * [ ] breeze-unstable 新的集合类型：`Tree`。
 * [ ] breeze-unstable 新的集合类型：`Table`。